pyproject.toml
[tool.poetry]
name = "impw"
version = "0.1.0"
description = ""
authors = ["marcelm9 <marcel.menzel09@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "impw", from = "main" }]

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.7.1"
pygame-ce = "^2.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
